Add copyright notices and break out UnknownUnitTypeException
